requirements.txt
# Core backend
Flask==3.0.3
Werkzeug==3.0.6
Jinja2==3.1.4
itsdangerous==2.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
MarkupSafe==3.0.2
requests==2.32.3
SQLAlchemy==2.0.29
python-dotenv==1.0.1
scapy==2.5.0
mlflow==2.11.3
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Data & ML
numpy==1.26.4
scipy==1.11.4
pandas==2.2.3
scikit-learn==1.3.2
joblib==1.4.2
threadpoolctl==3.5.0

# Database
psycopg2-binary==2.9.9

# Time & utils
python-dateutil>=2.9.0
pytz>=2024.1
tzdata>=2024.1
six>=1.16.0

# Testing
pytest>=7.0.0
